Misc changes.
Add new routes to known lists right a new service is added, so that a WD restart isn't required to add a new service. Also start new services at the same time, for the same reason, and stop them on deletion. Deleting a service now also deletes all the uploaded files and relate folders/extra crap belonging to that service.