Make Vault access by hosted services less of a tedious affair.
One long time annoyance I've had with using Vault is that whenever
the server restarts, I have to login to it, and edit config files to
manually set a vault token for each service that used it. Granted till
now that was only 1, but now that there's 2 (and a 3rd in development) a
better way was needed. This CL contains that better way. Frost itself is
now the only one that needs a manually set token, and that token can be
set via the ICAPI. Every other service can use Frost-generated AppRole
credentials to login to Vault and generate their tokens. This is
achieved by setting 3 environment vars on each service that contain the
Role ID, the Vault server endpoint and a temporary token that can only
be used to grab the Secret ID. The service-in-question will take that
info and login with it which will produce the token I was manually
setting before, automatically.

Also while I'm at, remove "dbcreds" and "vaulttoken" cases from ICAPI as
they no long make sense, and update "new-service" guide.